Restore: 'Common' replaced by 'Util' in comments
