+++ src/apev2/APEv2TagMapper.ts
@@ -1,9 +1,10 @@
-import {INativeTagMap} from "../tagmap";
+import {INativeTagMap} from "../common/GenericTagTypes";
+import {CommonTagMapper} from "../common/GenericTagMapper";
 
 /**
  * ID3v2.2 tag mappings
  */
-export const APEv2TagMap: INativeTagMap = {
+const apev2TagMap: INativeTagMap = {
   Title: 'title',
   Artist: 'artist',
   Artists: 'artists',
@@ -72,3 +73,26 @@ export const APEv2TagMap: INativeTagMap = {
   MUSICIP_PUID: 'musicip_puid',
   Weblink: 'website'
 };
+
+export class APEv2TagMapper extends CommonTagMapper {
+
+  public constructor() {
+
+    const upperCaseMap: INativeTagMap = {};
+
+    for (const tag in apev2TagMap) {
+      upperCaseMap[tag.toUpperCase()] = apev2TagMap[tag];
+    }
+
+    super(['APEv2'], upperCaseMap);
+  }
+
+  /**
+   * @tag  Native header tag
+   * @return common tag name (alias)
+   */
+  protected getCommonName(tag: string) {
+    return this.tagMap[tag.toUpperCase()];
+  }
+
+}
