+++ src/id3v2/ID3v2.ts
@@ -1,4 +1,4 @@
-import common from '../common/Common';
+import common from '../common/Util';
 import * as Token from "token-types";
 
 /**
