+++ src/ogg/OggParser.ts
@@ -1,5 +1,5 @@
 'use strict';
-import common from "../common/Common";
+import common from "../common/Util";
 import {ITokenParser} from "../ParserFactory";
 import * as strtok3 from "strtok3";
 import {INativeAudioMetadata, IOptions} from "../index";
