+++ src/riff/RiffParser.ts
@@ -7,7 +7,7 @@ import * as WaveChunk from "./../wav/WaveChunk";
 import {Readable} from "stream";
 import {ID3v2Parser} from "../id3v2/ID3v2Parser";
 import {IChunkHeader} from "../aiff/Chunk";
-import Common from "../common/Common";
+import Common from "../common/Util";
 import {FourCcToken} from "../common/FourCC";
 import {Promise} from "es6-promise";
