+++ test/test-picard-parsing.ts
@@ -2,8 +2,8 @@ import {} from "mocha";
 import {assert} from 'chai';
 import * as mm from '../src';
 import * as path from 'path';
-import {TagType} from "../src/tagmap";
 import * as crypto from "crypto";
+import {TagType} from "../src/common/GenericTagTypes";
 
 const t = assert;
